requirements.txt
# Core LangChain and LangGraph dependencies
langgraph==0.2.50
langchain==0.3.12
langchain-openai==0.2.14
langchain-core==0.3.26
langchain-community==0.3.11

# Database and data processing
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==1.26.4
sqlalchemy==2.0.36

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.6.1

# Additional utilities
requests==2.32.3
typing-extensions==4.12.2

# Development and debugging (optional)
black==24.10.0
flake8==7.1.1
pytest==8.3.4
